requirements.txt
appnope==0.1.0
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bleach==2.1.1
boto==2.48.0
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
decorator==4.1.2
docker-py==1.10.3
docker-pycreds==0.2.1
entrypoints==0.2.3
enum34==1.1.6
fastzbarlight==0.0.14
funcsigs==1.0.2
functools32==3.2.3.post2
gensim==2.3.0
go-vncdriver==0.4.19
html5lib==1.0b10
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.3
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.4
notebook==5.2.0
numpy==1.13.1
olefile==0.44
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
PyAutoGUI==0.9.36
pyglet==1.2.4
Pygments==2.2.0
PyMsgBox==1.0.6
pyobjc==4.0
pyobjc-core==4.0
pyobjc-framework-Accounts==4.0
pyobjc-framework-AddressBook==4.0
pyobjc-framework-AppleScriptKit==4.0
pyobjc-framework-AppleScriptObjC==4.0
pyobjc-framework-ApplicationServices==4.0
pyobjc-framework-Automator==4.0
pyobjc-framework-AVFoundation==4.0
pyobjc-framework-AVKit==4.0
pyobjc-framework-CalendarStore==4.0
pyobjc-framework-CFNetwork==4.0
pyobjc-framework-CloudKit==4.0
pyobjc-framework-Cocoa==4.0
pyobjc-framework-Collaboration==4.0
pyobjc-framework-Contacts==4.0
pyobjc-framework-ContactsUI==4.0
pyobjc-framework-CoreBluetooth==4.0
pyobjc-framework-CoreData==4.0
pyobjc-framework-CoreLocation==4.0
pyobjc-framework-CoreServices==4.0
pyobjc-framework-CoreText==4.0
pyobjc-framework-CoreWLAN==4.0
pyobjc-framework-CryptoTokenKit==4.0
pyobjc-framework-DictionaryServices==4.0
pyobjc-framework-DiskArbitration==4.0
pyobjc-framework-EventKit==4.0
pyobjc-framework-ExceptionHandling==4.0
pyobjc-framework-FinderSync==4.0
pyobjc-framework-FSEvents==4.0
pyobjc-framework-GameCenter==4.0
pyobjc-framework-GameController==4.0
pyobjc-framework-GameKit==4.0
pyobjc-framework-GameplayKit==4.0
pyobjc-framework-ImageCaptureCore==4.0
pyobjc-framework-IMServicePlugIn==4.0
pyobjc-framework-InputMethodKit==4.0
pyobjc-framework-InstallerPlugins==4.0
pyobjc-framework-InstantMessage==4.0
pyobjc-framework-Intents==4.0
pyobjc-framework-IOSurface==4.0
pyobjc-framework-iTunesLibrary==4.0
pyobjc-framework-LatentSemanticMapping==4.0
pyobjc-framework-LaunchServices==4.0
pyobjc-framework-LocalAuthentication==4.0
pyobjc-framework-MapKit==4.0
pyobjc-framework-MediaAccessibility==4.0
pyobjc-framework-MediaLibrary==4.0
pyobjc-framework-MediaPlayer==4.0
pyobjc-framework-ModelIO==4.0
pyobjc-framework-MultipeerConnectivity==4.0
pyobjc-framework-NetFS==4.0
pyobjc-framework-NetworkExtension==4.0
pyobjc-framework-NotificationCenter==4.0
pyobjc-framework-OpenDirectory==4.0
pyobjc-framework-Photos==4.0
pyobjc-framework-PhotosUI==4.0
pyobjc-framework-PreferencePanes==4.0
pyobjc-framework-PubSub==4.0
pyobjc-framework-QTKit==4.0
pyobjc-framework-Quartz==4.0
pyobjc-framework-SafariServices==4.0
pyobjc-framework-SceneKit==4.0
pyobjc-framework-ScreenSaver==4.0
pyobjc-framework-ScriptingBridge==4.0
pyobjc-framework-SearchKit==4.0
pyobjc-framework-ServiceManagement==4.0
pyobjc-framework-Social==4.0
pyobjc-framework-SpriteKit==4.0
pyobjc-framework-StoreKit==4.0
pyobjc-framework-SyncServices==4.0
pyobjc-framework-SystemConfiguration==4.0
pyobjc-framework-WebKit==4.0
PyScreeze==0.1.13
python-dateutil==2.6.1
PyTweening==1.0.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.10.0
scandir==1.6
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.3
stevedore==1.26.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
ujson==1.35
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
widgetsnbextension==3.0.6
zope.interface==4.4.3
